Update logtracer.py
Forced all data to be floats as previous behaviour was non-specific and this would lead to intermittent failures to write to influxdb since if the first instance sent is  an interger the field type gets set to interger then when we try to write a float influxdb returns an error. This resulted in gaps in data.